用户信息初始化，增加各类工具类，添加数据库sql日志

+++ src/main/java/com/chao/rabbitSolu/dao/CustomerInfDao.java
@@ -1,6 +1,9 @@
 package com.chao.rabbitSolu.dao;
 
+import com.chao.rabbitSolu.config.RabbitConfig;
 import com.chao.rabbitSolu.model.CustomerInf;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -12,5 +15,5 @@ import org.springframework.stereotype.Repository;
  * @Version 0.1
  **/
 @Repository
-public interface CustomerInfDao extends JpaRepository<CustomerInf,Integer> {
+public interface CustomerInfDao extends JpaRepository<CustomerInf,Long> {
 }
